requirements.txt
selenium==4.8.2
uvicorn==0.21.1
beautifulsoup4==4.11.2
fastapi==0.94.1
python-dotenv==1.0.0
python-decouple==3.8
webdriver-manager==3.8.5